feat: add line chart
