+++ src/routes/Interest/intereststock/chart/CandleChart.jsx
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect, useState } from "react";
 import ApexChart from "react-apexcharts";
 import { fetchStockPrice } from "../../../../lib/apis/stock";
 
-export default function SampleChart({ mode, stockKey, price }) {
+export default function CandleChart({ mode, stockKey, price }) {
   const [data, setData] = useState([]);
   const [chartType, setChartType] = useState("line");
   const [noContent, setNoContent] = useState(false);
@@ -29,24 +29,26 @@ export default function SampleChart({ mode, stockKey, price }) {
 
       if (response.result.length > 1) {
         const newList = response.result?.map((elem, index) => {
+          const timeStr = elem.localDate ? elem.localDate : elem.localDateTime;
           return {
-            x: elem.localDate ? elem.localDate : elem.localDateTime,
+            // x: elem.localDate ? elem.localDate : elem.localDateTime,
+            x: timeStrToTimestamp(timeStr),
             y: [
               elem.openPrice,
               elem.highPrice,
               elem.lowPrice,
-              elem.closePrice ? elem.closePrice : elem.currentPrice,
+              elem.closePrice
+                ? elem.closePrice
+                : Number.parseInt(elem.currentPrice),
             ],
           };
           // return elem.closePrice;
         });
-        console.log(newList);
         setData(newList);
         // console.log(newList[1].x - newList[0].x);
         setChartInterval(newList[1].x - newList[0].x);
-        setChartInterval(100);
       } else {
-        // setNoContent(true);
+        setNoContent(true);
       }
     }
     if (mode === "day") setChartType("candlestick");
@@ -79,10 +81,14 @@ export default function SampleChart({ mode, stockKey, price }) {
           const newList = data.filter((elem) => {
             return true;
           });
-          console.log(response);
+          // console.log(response);
           const addedList = response.result?.map((elem, index) => {
+            const timeStr = elem.localDate
+              ? elem.localDate
+              : elem.localDateTime;
             return {
-              x: elem.localDate ? elem.localDate : elem.localDateTime,
+              // x: elem.localDate ? elem.localDate : elem.localDateTime,
+              x: timeStrToTimestamp(timeStr),
               y: [
                 elem.openPrice,
                 elem.highPrice,
@@ -96,36 +102,28 @@ export default function SampleChart({ mode, stockKey, price }) {
           });
           newList.push(...addedList);
           setData(newList);
-          console.log(newList);
+          // console.log(newList);
           setUpdateData(false);
         }
       );
     }
   }, [updateData]);
 
-  // useEffect(() => {
-  //   if (noContent === true && mode === "day") {
-  //     console.log("No content");
-  //     setNoContent(false);
-  //     const str = getPrevDateStrings(prevDateCnt);
-  //     console.log(str);
+  const timeStrToTimestamp = useCallback((str) => {
+    const year = str.slice(0, 4);
+    const month = str.slice(4, 6);
+    const day = str.slice(6, 8);
+    if (str.length > 11) {
+      const hour = str.slice(8, 10);
+      const min = str.slice(10, 12);
 
-  //     fetchStockPrice(stockKey, "minute", str[0], str[1])
-  //       .then((response) => {
-  //         console.log(response);
-  //         return response.result;
-  //       })
-  //       .then((result) => {
-  //         console.log(result);
-  //         if (result.length > 0) {
-  //           setData(result);
-  //         } else {
-  //           setPrevDateCnt(prevDateCnt + 1);
-  //           setNoContent(true);
-  //         }
-  //       });
-  //   }
-  // }, [noContent, prevDateCnt]);
+      // console.log(year, month, day, hour, min);
+      return new Date(year, month, day, hour, min);
+    } else {
+      // console.log(year, month, day);
+      return new Date(year, month, day);
+    }
+  }, []);
 
   useEffect(() => {
     if (mode === "day" && data.length > 0) {
@@ -143,20 +141,6 @@ export default function SampleChart({ mode, stockKey, price }) {
     }
   }, [price]);
 
-  const getPrevDateStrings = useCallback((cnt) => {
-    const date = new Date(Date.now() - cnt * 1000 * 3600 * 24);
-    const from = new Date(date);
-    const end = new Date(date);
-
-    from.setHours(7);
-    end.setHours(16);
-
-    const fromString = getTimeString(from);
-    const endString = getTimeString(end);
-
-    return [fromString, endString];
-  }, []);
-
   const getFromDate = useCallback(
     (now) => {
       const fromDate = new Date(now);
@@ -167,7 +151,7 @@ export default function SampleChart({ mode, stockKey, price }) {
         fromDate.setMinutes(0);
         fromDate.setSeconds(0);
         fromDate.setMilliseconds(0);
-      } else if (mode === "month") {
+      } else if (mode === "week") {
         fromDate.setMonth(fromDate.getMonth() - 1);
         interval = "day";
       } else if (mode === "3month") {
@@ -177,6 +161,7 @@ export default function SampleChart({ mode, stockKey, price }) {
         fromDate.setFullYear(fromDate.getFullYear() - 1);
         interval = "week";
       }
+
       return { interval, fromDate };
     },
     [mode]
@@ -193,7 +178,6 @@ export default function SampleChart({ mode, stockKey, price }) {
   }, []);
 
   return data?.length > 0 ? (
-
     <ApexChart
       type={chartType}
       series={[
@@ -209,22 +193,27 @@ export default function SampleChart({ mode, stockKey, price }) {
           height: 350,
           type: { chartType },
         },
-        colors: ["#FF0000"],
         xaxis: {
           tooltip: {
             enabled: true,
+            formatter: function (value, timestamp) {
+              return new Date(value).toISOString().split(".")[0];
+            },
           },
-          type: mode === "day" ? "date" : "date",
+          type: "datetime",
           labels: {
             show: false,
           },
         },
         yaxis: {
           tooltip: {
-            enabled: true,
+            enabled: false,
+            formatter: function (value, val1) {
+              console.log(value, val1);
+            },
           },
           labels: {
-            show: false,
+            show: true,
             formatter: function (value) {
               return value.toFixed(0);
             },
@@ -241,45 +230,9 @@ export default function SampleChart({ mode, stockKey, price }) {
             },
           },
         },
-        annotations: {
-          // 예시로 선을 추가합니다. 여기에 원하는 어노테이션을 추가할 수 있습니다.
-          xaxis: [
-            {
-              x: data[15]?.x,
-              strokeDashArray: 0,
-              borderColor: "#775DD0",
-              label: {
-                borderColor: "#775DD0",
-                style: {
-                  color: "#fff",
-                  background: "#775DD0",
-                },
-                text: "어노테이션",
-              },
-            },
-          },
-        },
-        // annotations: {
-        //   // 예시로 선을 추가합니다. 여기에 원하는 어노테이션을 추가할 수 있습니다.
-        //   yaxis: [
-        //     {
-        //       x: data[0]?.x,
-        //       strokeDashArray: 0,
-        //       borderColor: "#775DD0",
-        //       label: {
-        //         borderColor: "#775DD0",
-        //         style: {
-        //           color: "#fff",
-        //           background: "#775DD0",
-        //         },
-        //         text: data[0].y[0],
-        //       },
-        //     },
-        //   ],
-        // },
       }}
     />
   ) : (
-    <></>
+    <div style={{ height: "262px" }}></div>
   );
 }
